% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/dailyExtreme.R
\name{dailyExtreme}
\alias{dailyExtreme}
\title{Extracts extreme values from sub-daily raster time series.}
\usage{
dailyExtreme(variable, min = T, cores = 10, timedate, clcall = NULL)
}
\arguments{
\item{variable}{A \code{RasterBrick} or \code{RasterStack} object with values
on a sub-daily resolution.}

\item{min}{A logical value indicating if daily minima (\code{min = TRUE}) or
daily maxima (\code{min = FALSE}) should be extracted.}

\item{cores}{The number of cores to be used in parallel computing.}

\item{timedate}{A \code{POSIXct} vector containing the time
information for each band/layer of \code{variable}.}

\item{clcall}{A function passed to \code{\link[parallel]{clusterCall}}.}
}
\value{
A list with the first element representing the extracted extreme values of the
raster time series as a \code{RasterBrick} and
the second element a \code{POSIXct} vector containing the date information
(i.e. the days).
}
\description{
\code{dailyExtreme} extracts extreme values from raster
based time series
(\code{RasterBrick} or \code{RasterStack} object, see:
\code{\link[raster]{Raster-class}}) on a sub-daily basis.
}
\examples{
#
}
