% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/dailySums.R
\name{dailySums}
\alias{dailySums}
\title{Calculates cumulative precipitation raster time series.}
\usage{
dailySums(variable, cores = 10, timedate, clcall = NULL)
}
\arguments{
\item{variable}{A \code{RasterBrick} or \code{RasterStack} object with
values on a sub-daily resolution.}

\item{cores}{The number of cores to be used in parallel computing.}

\item{timedate}{A \code{POSIXct} vector containing the time
information for each band/layer of \code{variable}.}

\item{clcall}{A function passed to \code{\link[parallel]{clusterCall}}.}
}
\value{
A list with the first element representing the aggregated
raster time series with daily summed values as a \code{RasterBrick} and
the second element a \code{POSIXct} vector containing the date information
(i.e. the days).
}
\description{
\code{dailySums} computes total daily sums of a raster based time series
(\code{RasterBrick} or \code{RasterStack} object, see:
\code{\link[raster]{Raster-class}}) on a sub-daily basis by computing
the sum of each raster cell for a specified temporal resolution.
In principle, this function can be used in order to compute any
aggregate value that is based on sums.
}
\examples{
#
}
