% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/evaluateDroughtNema.R
\name{evaluateDroughtNema}
\alias{evaluateDroughtNema}
\title{Evaluates Drought Conditions Based on Raster Time Series.}
\usage{
evaluateDroughtNema(precipitation, ltmprecipitation, airtemperature,
  ltmairtemperature, ltsdairtemperature, dailymaxairtemperature, dailymeanrh,
  naturalzonesraster, resolution = "monthly", timedate_aggregated,
  timedate_daily, cores = 10, clcall = NULL)
}
\arguments{
\item{precipitation}{A \code{RasterBrick} or \code{RasterStack} object with
(mean) total precipitation values for each time interval within the target time
period. See the details section.}

\item{ltmprecipitation}{A \code{RasterBrick} or \code{RasterStack} object with
long-term mean total precipitation values for all time intervals denoted by
\code{resolution} within a year. See the details section.}

\item{airtemperature}{A \code{RasterBrick} or \code{RasterStack} object with
mean air temperature values for each time interval within the target time
period. See the details section.}

\item{ltmairtemperature}{A \code{RasterBrick} or \code{RasterStack} object with
long-term mean air temperature values for all time intervals denoted by
\code{resolution} within a year. See the details section.}

\item{ltsdairtemperature}{A \code{RasterBrick} or \code{RasterStack} object with
standard deviation values of mean air temperature values within the time interval
considered as long-term time interval. See the details section.}

\item{dailymaxairtemperature}{A \code{RasterBrick} or \code{RasterStack} object with
daily maximum air temperature values within the target time period. See the
details section.}

\item{dailymeanrh}{A \code{RasterBrick} or \code{RasterStack} object with
daily mean relative air humidity values within the target time period. See the
details section.}

\item{naturalzonesraster}{A \code{RasterLayer} object containing information
on landcover classes for which additional criteria concerning drought conditions are
defined. Classes that should not be considered must have the value "0". Mountainous
regions must have the value "1". Forest steppe and steppe must have the value "2".
The gobi-desert must have the value "3".}

\item{resolution}{A character value indicating the temporal resolution of the
\code{Raster*} time series data for the target time period (\code{precipitation} and
\code{airtemperature}). One of \code{"monthly"}, \code{"ftdi"} (fixed ten-day
intervals) or \code{"mwtdi"} (moving window ten-day intervals).}

\item{timedate_aggregated}{A \code{POSIXct} vector containing the time
information for each band/layer of \code{precipitation} and
\code{airtemperature}.}

\item{timedate_daily}{A \code{POSIXct} vector containing the time
information for each band/layer of \code{dailymaxairtemperature} and
\code{dailymeanrh}.}

\item{cores}{The number of cores to be used in parallel computing.}

\item{clcall}{A function passed to \code{\link[parallel]{clusterCall}}.}
}
\value{
A list with the first element representing daily values if drought (2), near
drought (1) or no drought (0) conditions prevail, as a \code{RasterBrick} object, and
the second element being equal to \code{timedate_daily} and indicating the time information
relating to the layers of the first element.
}
\description{
\code{evaluateDroughtNema} evaluates based on a set of weather variables as
raster based time series (\code{RasterBrick} or \code{RasterStack} object, if
for a given spatial and temporal point drought or near drought conditions are
prevalent or not (according to the classification of the NEMA).
}
\details{
The start and end time point of the target time interval are derived from
\code{timedate_daily}. All time intervals of \code{timedate_aggregated} are supposed
to fit perfectly within \code{timedate_daily} (i.e. there are no more or less days in
\code{timedate_daily}). All \code{Raster*} time series data are supposed to have the
same spatial extent an resolution. \code{precipitation} and \code{airtemperature} are
supposed to have the same temporal resolution. \code{dailymaxairtemperature} and
\code{dailymeanrh} are supposed to have the same temporal resolution.
}
\examples{
#
}
