% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/evaluateDroughtNemaAithumidity.R
\name{evaluateDroughtNemaAithumidity}
\alias{evaluateDroughtNemaAithumidity}
\title{Evaluates Drought Conditions Relating to Air Humidity Based on Raster Time Series.}
\usage{
evaluateDroughtNemaAithumidity(airhumidity, airhumidity_t)
}
\arguments{
\item{airhumidity}{A \code{RasterBrick} or \code{RasterStack} object with
mean relative air humidity values on a daily basis within the target time
period.}

\item{airhumidity_t}{A \code{POSIXct} vector containing the time information
for all layers in \code{airhumidity} as returned by \code{\link{dailyMeans}}
(each element denoting the respective day).}
}
\value{
A list with two elements:
\describe{
  \item{\code{droughtairhumidity}}{A \code{RasterBrick} object containing for each
  fixed ten-day interval in \code{airhumidity_t} a layer classifying if there were
  drought conditions according to the relative air humidity or not.}
  \item{\code{time}}{A character vector specifying the date of the first day of the
  respective fixed ten-day interval.}
}
}
\description{
\code{evaluateDroughtNemaAithumidity} evaluates based on relative air humidity
values for a target time period as
raster based time series (\code{RasterBrick} or \code{RasterStack} object, if
for a given spatial and temporal point drought or near drought conditions are
prevalent or not (according to the classification of the NEMA). This function
considers only the additional criteria related to the relative air humidity.
}
\examples{
#
}
