% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/evaluateDroughtNemaPrecipitation.R
\name{evaluateDroughtNemaPrecipitation}
\alias{evaluateDroughtNemaPrecipitation}
\title{Evaluates Drought Conditions Relating to Precipitation Based on Raster Time Series.}
\usage{
evaluateDroughtNemaPrecipitation(precipitation, ltmprecipitation,
  precipitation_t, ltmprecipitation_t)
}
\arguments{
\item{precipitation}{A \code{RasterBrick} or \code{RasterStack} object with
(mean) total precipitation values for each time interval within the target time
period. See the details section. Time intervals are supposed to represent fixed
ten-day intervals.}

\item{ltmprecipitation}{A \code{RasterBrick} or \code{RasterStack} object with
long-term mean total precipitation values for fixed ten-day intervals within a
year. See the details section.}

\item{precipitation_t}{A \code{POSIXct} vector containing the time information
for all layers in \code{precipitation} as returned by \code{\link{weatherMean}}
(each element denoting the first day of the respective ten-day interval).}

\item{ltmprecipitation_t}{A \code{POSIXct} vector containing the time information
for all layers in \code{ltmprecipitation} as returned by \code{\link{intervalMean}}
(each element denoting the first day of the respective ten-day interval).}
}
\description{
\code{evaluateDroughtNemaPrecipitation} evaluates based on precipitation values
for a target time period and representing long term mean values as
raster based time series (\code{RasterBrick} or \code{RasterStack} object, if
for a given spatial and temporal point drought or near drought conditions are
prevalent or not (according to the classification of the NEMA). This function
considers only criteria related to precipitation.
}
\details{
The start and end time point of the target time interval are derived from
\code{timedate_daily}. All time intervals of \code{timedate_aggregated} are supposed
to fit perfectly within \code{timedate_daily} (i.e. there are no more or less days in
\code{timedate_daily}). All \code{Raster*} time series data are supposed to have the
same spatial extent and resolution.
}
\examples{
#
}
