% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/evaluateDroughtNemaTemepratureAdditional.R
\name{evaluateDroughtNemaTemepratureAdditional}
\alias{evaluateDroughtNemaTemepratureAdditional}
\title{Evaluates Drought Conditions (Additional Criteria) Relating to Temperature Based on Raster Time Series.}
\usage{
evaluateDroughtNemaTemepratureAdditional(temperature, temperature_t, landcover)
}
\arguments{
\item{temperature}{A \code{RasterBrick} or \code{RasterStack} object with
(mean) temperature values on a daily basis within the target time
period.}

\item{temperature_t}{A \code{POSIXct} vector containing the time information
for all layers in \code{airhumidity} as returned by \code{\link{dailyMeans}}
(each element denoting the respective day).}

\item{landcover}{A \code{RasterBrick} object containing the following layers
in the following sequence
\describe{
  \item{\code{"alpine"}}{}
  \item{\code{"steppe"}}{}
  \item{\code{"forest_steppe"}}{}
  \item{\code{"desert"}}{}
  \item{\code{"desert_steppe"}}{}
  \item{\code{"lakes"}}{}
  \item{\code{"taiga"}}{}
}
Each raster layer has a value of \code{1} if the respective landcover class exists
for a pixel and \code{NA} if not.}
}
\value{
A list with two elements:
\describe{
  \item{\code{droughttemperatureadditional}}{A \code{RasterBrick} object containing for each
  fixed ten-day interval in \code{temperature_t} a layer classifying if there were
  drought conditions according to the relative air humidity or not.}
  \item{\code{time}}{A character vector specifying the date of the first day of the
  respective fixed ten-day interval.}
}
}
\description{
\code{evaluateDroughtNemaTemepratureAdditional} evaluates based on air temeprature
values for a target time period as
raster based time series (\code{RasterBrick} or \code{RasterStack} object, if
for a given spatial and temporal point drought or near drought conditions are
prevalent or not (according to the classification of the NEMA).
This function considers only additional criteria related to temerature. See
\code{\link{evaluateDroughtNemaTemepratureMain}} for temperature related
threshold values that represent main criteria.
}
\examples{
#
}
