% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/evaluateDroughtNemaTemepratureMain.R
\name{evaluateDroughtNemaTemepratureMain}
\alias{evaluateDroughtNemaTemepratureMain}
\title{Evaluates Drought Conditions (Main Criteria) Relating to Temperature Based on Raster Time Series.}
\usage{
evaluateDroughtNemaTemepratureMain(temperature, ltmtemperature, ltsdtemperature,
  temperature_t, ltmtemperature_t)
}
\arguments{
\item{temperature}{A \code{RasterBrick} or \code{RasterStack} object with
(mean) temperature values for each time interval within the target time
period. See the details section. Time intervals are supposed to represent fixed
ten-day intervals.}

\item{ltmtemperature}{A \code{RasterBrick} or \code{RasterStack} object with
long-term mean temperature values for fixed ten-day intervals within a
year. See the details section.}

\item{ltsdtemperature}{A \code{RasterBrick} or \code{RasterStack} object with
standard deviation values of mean air temperature values within the time interval
considered as long-term time interval. \code{ltsdairtemperature} must have the
same number of layers as \code{ltmtemperature}.}

\item{ltmtemperature_t}{A \code{POSIXct} vector containing the time information
for all layers in \code{ltmtemperature} as returned by \code{\link{intervalMean}}
(each element denoting the first day of the respective ten-day interval).}

\item{temeprature_t}{A \code{POSIXct} vector containing the time information
for all layers in \code{temperature} as returned by \code{\link{weatherMean}}
(each element denoting the first day of the respective ten-day interval).}
}
\description{
\code{evaluateDroughtNemaTemepratureMain} evaluates based on temperature values
for a target time period and representing long term mean values as
raster based time series (\code{RasterBrick} or \code{RasterStack} object, if
for a given spatial and temporal point drought or near drought conditions are
prevalent or not (according to the main criteria classification of the NEMA).
This function considers only main criteria related to temerature. See
\code{\link{evaluateDroughtNemaTemepratureAdditional}} for temperature related
threshold values that represent additional criteria.
}
\details{
The start and end time point of the target time interval are derived from
\code{timedate_daily}. All time intervals of \code{timedate_aggregated} are supposed
to fit perfectly within \code{timedate_daily} (i.e. there are no more or less days in
\code{timedate_daily}). All \code{Raster*} time series data are supposed to have the
same spatial extent and resolution.
}
\examples{
#
}
