% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/monthlyLongTermMeans.R
\name{monthlyLongTermMeans}
\alias{monthlyLongTermMeans}
\title{Computes long-term monthly means of raster time series.}
\usage{
monthlyLongTermMeans(variable, tstart, tend, cores = 12, timedate,
  clcall = NULL)
}
\arguments{
\item{variable}{A \code{RasterBrick} or \code{RasterStack} object with
values on a sub-monthly resolution.}

\item{tstart}{Numeric value (four places) for the first year of the time period
to consider as long-term interval, e.g. \code{tstart = 1984}.}

\item{tend}{Numeric value (four places) for the last year of the time period
to consider as long-term interval, e.g. \code{tstart = 2014}.}

\item{cores}{The number of cores to be used in parallel computing.}

\item{timedate}{A \code{POSIXct} vector containing the time
information for each band/layer of \code{variable}.}

\item{clcall}{A function passed to \code{\link[parallel]{clusterCall}}.}
}
\value{
A list with the first element representing the aggregated
raster time series with daily mean values as a \code{RasterBrick} and
the second element a \code{POSIXct} vector containing the date information
(i.e. the days).
}
\description{
\code{monthlyLongTermMeans} aggregates raster based time series
(\code{RasterBrick} or \code{RasterStack} object, see:
\code{\link[raster]{Raster-class}}) by computing the long-term
mean values for a specified time period grouped by months.
}
\examples{
#
}
