% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/relErainterimAirhumidity.R
\name{relErainterimAirhumidity}
\alias{relErainterimAirhumidity}
\title{Computes the relative air humidity for raster based time series.}
\usage{
relErainterimAirhumidity(dt, t, interval = NULL, cores = 10, timedate,
  clcall = NULL)
}
\arguments{
\item{dt}{A \code{RasterBrick} or \code{RasterStack} object with values
for the dew point temperature. \code{dt} is supposed to have the same temporal extent and
resolution as \code{t}.}

\item{t}{A \code{RasterBrick} or \code{RasterStack} object with values for the air
temperature. \code{t} is supposed to have the same temporal extent and resolution
as \code{dt}.}

\item{cores}{The number of cores to be used in parallel computing.}

\item{timedate}{A \code{POSIXct} vector containing the time
information for each band/layer of \code{variable}.}

\item{clcall}{A function passed to \code{\link[parallel]{clusterCall}}.}
}
\value{
A list with the first element representing the relative air humidity [%]
time series with as a \code{RasterBrick} with the same temporal resultution as
\code{dt} and \code{t} and
the second element a \code{POSIXct} vector containing the time information.
}
\description{
\code{relErainterimAirhumidity} calculates the relative air humidity [%] of raster based time series
(\code{RasterBrick} or \code{RasterStack} object, see:
\code{\link[raster]{Raster-class}}) derived from netcdf raster files as downloaded from
ERA interim as suggested by
\url{https://software.ecmwf.int/wiki/display/CKB/Do+ERA+datasets+contain+parameters+for+near-surface+humidity},
based on the dew point temperature and the air temperature. This is done using thw Teten's
formula for saturation over water as suggested at
\url{https://software.ecmwf.int/wiki/display/CKB/Do+ERA+datasets+contain+parameters+for+near-surface+humidity} and in
\url{https://www.ecmwf.int/sites/default/files/elibrary/2016/16648-part-iv-physical-processes.pdf}.
}
\examples{
#
}
