% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/spatAggregate.R
\name{spatAggregate}
\alias{spatAggregate}
\title{Spatially aggregates raster time series within geometries.}
\usage{
spatAggregate(variable, spatialentity, tstart, tend, fn = "mean")
}
\arguments{
\item{variable}{A \code{RasterBrick} or \code{RasterStack} object.}

\item{spatialentity}{A \code{SpatialPolygonsDataFrame} of the  spatial entity to
iaggregate within..}

\item{tstart}{Start of the time interval during which to perform aggregation
given as character values in the format "yyyy-mm-dd".}

\item{tend}{End of the time interval during which to perform aggregation given
as character values in the format "yyyy-mm-dd".}

\item{fn}{Function to be used for aggregation. One of \code{mean}, \code{min}
or \code{max}.}
}
\value{
A data frame containing the aggregated vaules of each raster layer (time instance)
within each spatial entity
}
\description{
\code{spatAggregate} performs aggregation of raster based time series
(\code{RasterBrick} or \code{RasterStack} object, see:
\code{\link[raster]{Raster-class}}) within certain spatial entity (polygon geometry)
 applying \code{over} from package \code{sp} on all raster layers (time instances)
 within a given time interval.
}
\examples{
#
}
