% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/tenDayLongTermMWMeans.R
\name{tenDayLongTermMWMeans}
\alias{tenDayLongTermMWMeans}
\title{Computes long-term ten-day moving window means of raster time series.}
\usage{
tenDayLongTermMWMeans(variable, tstart, tend, cores = 10, timedate,
  clcall = NULL)
}
\arguments{
\item{variable}{A \code{RasterBrick} or \code{RasterStack} object with
values on a daily resolution.}

\item{tstart}{Numeric value (four places) for the first year of the time period
to consider as long-term interval, e.g. \code{tstart = 1984}.}

\item{tend}{Numeric value (four places) for the last year of the time period
to consider as long-term interval, e.g. \code{tstart = 2014}.}

\item{cores}{The number of cores to be used in parallel computing.}

\item{timedate}{A \code{POSIXct} vector containing the time
information for each band/layer of \code{variable}.}

\item{clcall}{A function passed to \code{\link[parallel]{clusterCall}}.}
}
\value{
A list with the first element representing the aggregated
raster time series with moving window ten-day mean values as a \code{RasterBrick} and
the second element a \code{POSIXct} vector containing the date information
(i.e. the days representing the start of each moving window interval).
}
\description{
\code{tenDayLongTermMWMeans} aggregates raster based time series
(\code{RasterBrick} or \code{RasterStack} object, see:
\code{\link[raster]{Raster-class}}) by computing the long-term
mean values for a specified time period grouped by ten-day intervals.
Ten-day intervals are constructed for each day of a year by sliding a
moving window with a length of 10 days along the sequence of days of a
year. For leap years, the corresponding ten day interval containing the
29 th February. In all other years, the interval reaches one day more
in the future and the ten day interval for the 29 th February is the
same as for the 28 th february.
}
\examples{
#
}
