% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/timeZoneConvert.R
\name{timeZoneConvert}
\alias{timeZoneConvert}
\title{Converts the time zone of a date and time vector.}
\usage{
timeZoneConvert(variable, timezone = "Asia/Ulan_Bator")
}
\arguments{
\item{variable}{A \code{character} vector convertable to a \code{POSIXct} object.}

\item{timezone}{A character value representing a timezone
(\code{\link[base]{timezones}}).}
}
\value{
A \code{POSIXct} vector with converted time and date information.
}
\description{
\code{timeZoneConvert} converts date and time information from a \code{character}
vector convertable to a \code{POSIXct} object to a user specified timezone.
}
\examples{
#
}
