% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/weatherMean.R
\name{weatherMean}
\alias{weatherMean}
\title{Aggregates daily raster time series.}
\usage{
weatherMean(variable, tstart = NULL, tend = NULL, resolution = "monthly",
  timedate, cores = 10, clcall = NULL)
}
\arguments{
\item{variable}{A \code{RasterBrick} or \code{RasterStack} object with
values on a daily resolution.}

\item{tstart}{Numeric value (four places) for the first year of the time period
to consider as long-term interval, e.g. \code{tstart = 1984}.}

\item{tend}{Numeric value (four places) for the last year of the time period
to consider as long-term interval, e.g. \code{tstart = 2014}.}

\item{resolution}{Character value indicating the temporal resolution of the mean values.
If \code{resolution = "fixedtendays"}, mean values will be computed for each
fixed ten-day interval as defined in \code{\link{intervalMean}}. If
\code{resolution = "mwtendays"}, mean values will be computed for each
moving window ten-day interval as defined  in
\code{\link{tenDayLongTermMWMeans}}. If \code{resolution = "months"}, mean
values will be computed for each month.}

\item{timedate}{A \code{POSIXct} vector containing the time
information for each band/layer of \code{variable}.}

\item{cores}{The number of cores to be used in parallel computing.}

\item{clcall}{A function passed to \code{\link[parallel]{clusterCall}}.}
}
\value{
A list with the first element representing the aggregated
raster time series with mean values as specified by \code{resolution} as
a \code{RasterBrick} and the second element a \code{POSIXct} vector containing
information on the aggregated time intervals as specified by \code{resolution}.
For fixed ten day intervals, this is an integer vector with increasing values for
each ten-day interval.
}
\description{
\code{weatherMean} aggregates raster based time series
(\code{RasterBrick} or \code{RasterStack} object, see:
\code{\link[raster]{Raster-class}}) by computing mean values
along the time series (i.e. no long-term mean values) as
monthly mean values, moving window ten-day interval mean
values as defined in \code{\link{tenDayLongTermMWMeans}} and
fixed ten-day interval mean values as defined in
\code{\link{intervalMean}}.
}
\examples{
#
}
