% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/weatherSD.R
\name{weatherSD}
\alias{weatherSD}
\title{Computes the standard deviation of raster time series.}
\usage{
weatherSD(variable, tstart = NULL, tend = NULL, resolution = "monthly",
  timedate, cores = 10, clcall = NULL)
}
\arguments{
\item{variable}{A \code{RasterBrick} or \code{RasterStack} object with
values on a daily resolution.}

\item{timedate}{A \code{POSIXct} vector containing the time
information for each band/layer of \code{variable}.}

\item{cores}{The number of cores to be used in parallel computing.}

\item{clcall}{A function passed to \code{\link[parallel]{clusterCall}}.}
}
\value{
A list with the first element representing the aggregated
raster time series with mean values as specified by \code{resolution} as
a \code{RasterBrick} and the second element a \code{POSIXct} vector containing
information on the aggregated time intervals as specified by \code{resolution}.
For fixed ten day intervals, this is an integer vector with increasing values for
each ten-day interval.
}
\description{
\code{weatherSD} computes the standard deviation of raster
based time series (\code{RasterBrick} or \code{RasterStack} object, see:
\code{\link[raster]{Raster-class}}) for the respective temporal
resolution.
}
\examples{
#
}
