% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/windSpeedErainterim.R
\name{windSpeedErainterim}
\alias{windSpeedErainterim}
\title{Computes total wind speed for raster based time series.}
\usage{
windSpeedErainterim(u10, v10, cores = 10, timedate, clcall = NULL)
}
\arguments{
\item{u10}{A \code{RasterBrick} or \code{RasterStack} object with values
for the wind speed at 10 m height in east direction (negative signs indicate
a wind direction to the west). \code{u10} is
supposed to have the same temporal extent and resolution as \code{v10}.}

\item{v10}{A \code{RasterBrick} or \code{RasterStack} object with values
for the wind speed at 10 m height in north direction (negative signs indicate
a wind direction to the south). \code{v10} is
supposed to have the same temporal extent and resolution as \code{u10}.}

\item{cores}{The number of cores to be used in parallel computing.}

\item{timedate}{A \code{POSIXct} vector containing the time
information for each band/layer of \code{variable}.}

\item{clcall}{A function passed to \code{\link[parallel]{clusterCall}}.}
}
\value{
A list with the first element representing the relative air humidity [%]
time series with as a \code{RasterBrick} with the same temporal resultution as
\code{dt} and \code{t} and
the second element a \code{POSIXct} vector containing the time information.
}
\description{
\code{windSpeedErainterim} calculates the total wind speed [m/s] of raster based time series
(\code{RasterBrick} or \code{RasterStack} object, see:
\code{\link[raster]{Raster-class}}) derived from netcdf raster files as downloaded from
ERA interim. This is done by computing the length of the total vector from the north-south
and east-west components as suggested by
\url{https://confluence.ecmwf.int/pages/viewpage.action?pageId=77229867}.
}
\examples{
#
}
